requirements.txt
colormath
pygame
pillow
